Refactor state of the game of life object
- It makes sense for the game of life (gol) object to keep more information
  about the system. Specifically:
  - The total grid size that it contributes to.
  - The total number of processes to whom the grid is shared to.
  - The rank of the process that handles the specific gol.
  - The MPI communicator to use to communicate with neighboring processe.
- This way function interfaces is kept minimal and all stuff is internal,
  or implementation-dependent.

Notes:

It is restrictive in the way that you cannot use other communicator,
different amount of processes, or different process responsible for a gol.
Those limitations apply for a specific grid of Game of Life.

However, those restrictions are reasonable, since for the processes's responsibility
or number, or the communicator to change you need to redistribute the gols processes
all over again. There is no interfae or need to do this, so the eternal state of the
gols does not create any problems.